feature/settings: fixing error regarding business account not having a profile picture selected during the reghistration

+++ lib/src/homepage.dart
@@ -16,7 +16,6 @@ import 'navbar.dart';
 import 'radar/radar.dart';
 import 'settings/settings.dart';
 
-
 class HomePage extends StatefulWidget {
   const HomePage({ Key? key}) : super(key: key);
 
@@ -64,37 +63,43 @@ class _HomePage extends State<HomePage> {
   /// Preloads the profile picture from Firebase Storage
   Future<dynamic> getImageFile() async {
     if (accountData!.isNotEmpty) {
-      String imageName = accountData!['profile_picture'];
-      final tempDir = await getTemporaryDirectory();
-      final File tempFile = File('${tempDir.path}/$imageName');
+      String? imageName = accountData!['profile_picture'];
 
-      /// If file doesn't exist, it will try downloading
-      if (!tempFile.existsSync()) {
-        try {
-          tempFile.create(recursive: true);
-          await firebase_storage.FirebaseStorage.instance.ref(
-              '/profile_pictures/$imageName').writeToFile(tempFile);
-          setState(() {
-            _tempImageFile = tempFile;
-            menuSelects = <Widget>[
-              Discover(isBusiness: accountData!['business']),
-              accountData!['business'] == true ? EditEntityPage(userData: accountData!, profilePicture: _tempImageFile) : Radar(),
-              const Favorites(),
-              AppSettings(userData: accountData!, profilePicture: _tempImageFile)
-            ];
-          });
-        } catch (e) {
-          /// If there is an error the created file will be deleted
-          await tempFile.delete(recursive: true);
+      if (imageName != null) {
+        final tempDir = await getTemporaryDirectory();
+        final File tempFile = File('${tempDir.path}/$imageName');
+
+        /// If file doesn't exist, it will try downloading
+        if (!tempFile.existsSync()) {
+          try {
+            tempFile.create(recursive: true);
+            await firebase_storage.FirebaseStorage.instance.ref(
+                '/profile_pictures/$imageName').writeToFile(tempFile);
+            setState(() {
+              _tempImageFile = tempFile;
+              menuSelects = <Widget>[
+                Discover(isBusiness: accountData!['business']),
+                accountData!['business'] == true ? EditEntityPage(
+                    userData: accountData!) : Radar(),
+                const Favorites(),
+                AppSettings(
+                    userData: accountData!, profilePicture: _tempImageFile)
+              ];
+            });
+          } catch (e) {
+            /// If there is an error the created file will be deleted
+            await tempFile.delete(recursive: true);
+          }
         }
       } else {
         setState(() {
-          _tempImageFile = tempFile;
           menuSelects = <Widget>[
             Discover(isBusiness: accountData!['business']),
-            accountData!['business'] == true ? EditEntityPage(userData: accountData!, profilePicture: _tempImageFile) : Radar(), //TODO: FAVORITES
+            accountData!['business'] == true ? EditEntityPage(
+                userData: accountData!) : Radar(), //TODO: FAVORITES
             const Favorites(),
-            AppSettings(userData: accountData!, profilePicture: _tempImageFile)
+            AppSettings(
+                userData: accountData!, profilePicture: null)
           ];
         });
       }
